+++ birdo/src/birdo/utilities/enemy.java
@@ -249,8 +249,8 @@ public class enemy extends player {
 		
 		if (type == "laser") {
 			feather f = new feather (this.x, this.y, true);
-			f.dx = -1 * (int) (5 * Math.cos(shotmultiplier1 * Math.PI / 180));
-			f.dy = -1 * (int) (5 * Math.sin(shotmultiplier1 * Math.PI / 180));
+			f.dx = -1 * (5 * Math.cos(shotmultiplier1 * Math.PI / 180));
+			f.dy = -1 * (5 * Math.sin(shotmultiplier1 * Math.PI / 180));
 			shotmultiplier1--;
 			if (shotmultiplier1 <= -45)
 				shotmultiplier1 = 45;
